+++ native/async_encoder.h
@@ -71,6 +71,11 @@ private:
 
     // Worker thread
     std::thread workerThread_;
+    bool envTeardown_ = false;
+    Napi::ThreadSafeFunction tsfnJobDone_;
+    int activeJobs_ = 0;
+    void JobSubmitted(Napi::Env env);
+    void JobFinished(Napi::Env env);
     std::atomic<bool> running_{false};
     std::atomic<bool> configured_{false};